app/requirements.txt
alembic==1.13.2
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
geographiclib==2.0
geopy==2.4.1
greenlet==3.1.0
idna==3.10
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2==2.9.9
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.2.3
